+++ lib/slack/web/api/errors.rb
@@ -178,6 +178,7 @@ module Slack
         class ErrorBadWide < SlackError; end
         class ErrorInvalidAlias < SlackError; end
         class ErrorInvalidEmojiToRename < SlackError; end
+        class ErrorLowerCaseNamesOnly < SlackError; end
         class ErrorMissingName < SlackError; end
         class ErrorNameTaken < SlackError; end
         class ErrorNameTakenI18n < SlackError; end
@@ -207,6 +208,7 @@ module Slack
         class FailedToAliasEmoji < SlackError; end
         class FailedToFetchActiveTeam < SlackError; end
         class FailedToFetchInfo < SlackError; end
+        class FailedToRemoveEmoji < SlackError; end
         class FailedToRemoveGroupUsers < SlackError; end
         class FailedToRemoveUserFromWorkspace < SlackError; end
         class FailedToRenameEmoji < SlackError; end
@@ -336,6 +338,7 @@ module Slack
         class InvalidScopes < SlackError; end
         class InvalidSearchChannelType < SlackError; end
         class InvalidSetting < SlackError; end
+        class InvalidShortcutType < SlackError; end
         class InvalidSort < SlackError; end
         class InvalidSortDir < SlackError; end
         class InvalidSource < SlackError; end
@@ -367,6 +370,9 @@ module Slack
         class InvalidVisibility < SlackError; end
         class InvalidWorkflowAuthStrategy < SlackError; end
         class InvalidWorkspaceFilter < SlackError; end
+        class InviteAlreadyApproved < SlackError; end
+        class InviteAlreadyDenied < SlackError; end
+        class InviteExpired < SlackError; end
         class InviteFromSameOrg < SlackError; end
         class InviteLookupError < SlackError; end
         class InviteNotFound < SlackError; end
@@ -537,11 +543,11 @@ module Slack
         class RestrictedActionReadOnlyChannel < SlackError; end
         class RestrictedActionThreadLocked < SlackError; end
         class RestrictedActionThreadOnlyChannel < SlackError; end
+        class RestrictedPlanLevel < SlackError; end
         class RestrictedTooMany < SlackError; end
         class RetentionOverrideNotAllowed < SlackError; end
         class SentRecently < SlackError; end
         class ServiceUnavailable < SlackError; end
-        class SessionExpired < SlackError; end
         class SessionInvalidationFailed < SlackError; end
         class SessionNotFound < SlackError; end
         class SessionResetNotAllowed < SlackError; end
@@ -570,6 +576,7 @@ module Slack
         class TeamNotFound < SlackError; end
         class TeamNotOnEnterprise < SlackError; end
         class TeamQuotaExceeded < SlackError; end
+        class TeamTierCannotCreateChannelCanvases < SlackError; end
         class TeamsNotFound < SlackError; end
         class TemplateNotVisible < SlackError; end
         class ThreadLocked < SlackError; end
@@ -634,9 +641,11 @@ module Slack
         class UserAlreadyDeleted < SlackError; end
         class UserAlreadyTeamMember < SlackError; end
         class UserCannotCreateChannel < SlackError; end
+        class UserCannotManagePublicChannels < SlackError; end
         class UserCannotManageWorkspace < SlackError; end
         class UserDisabled < SlackError; end
         class UserDoesNotOwnChannel < SlackError; end
+        class UserEmailUnverified < SlackError; end
         class UserIsAlreadyDeleted < SlackError; end
         class UserIsBot < SlackError; end
         class UserIsNotAGuest < SlackError; end
@@ -833,6 +842,7 @@ module Slack
           'error_bad_wide' => ErrorBadWide,
           'error_invalid_alias' => ErrorInvalidAlias,
           'error_invalid_emoji_to_rename' => ErrorInvalidEmojiToRename,
+          'error_lower_case_names_only' => ErrorLowerCaseNamesOnly,
           'error_missing_name' => ErrorMissingName,
           'error_name_taken' => ErrorNameTaken,
           'error_name_taken_i18n' => ErrorNameTakenI18n,
@@ -862,6 +872,7 @@ module Slack
           'failed_to_alias_emoji' => FailedToAliasEmoji,
           'failed_to_fetch_active_team' => FailedToFetchActiveTeam,
           'failed_to_fetch_info' => FailedToFetchInfo,
+          'failed_to_remove_emoji' => FailedToRemoveEmoji,
           'failed_to_remove_group_users' => FailedToRemoveGroupUsers,
           'failed_to_remove_user_from_workspace' => FailedToRemoveUserFromWorkspace,
           'failed_to_rename_emoji' => FailedToRenameEmoji,
@@ -991,6 +1002,7 @@ module Slack
           'invalid_scopes' => InvalidScopes,
           'invalid_search_channel_type' => InvalidSearchChannelType,
           'invalid_setting' => InvalidSetting,
+          'invalid_shortcut_type' => InvalidShortcutType,
           'invalid_sort' => InvalidSort,
           'invalid_sort_dir' => InvalidSortDir,
           'invalid_source' => InvalidSource,
@@ -1022,6 +1034,9 @@ module Slack
           'invalid_visibility' => InvalidVisibility,
           'invalid_workflow_auth_strategy' => InvalidWorkflowAuthStrategy,
           'invalid_workspace_filter' => InvalidWorkspaceFilter,
+          'invite_already_approved' => InviteAlreadyApproved,
+          'invite_already_denied' => InviteAlreadyDenied,
+          'invite_expired' => InviteExpired,
           'invite_from_same_org' => InviteFromSameOrg,
           'invite_lookup_error' => InviteLookupError,
           'invite_not_found' => InviteNotFound,
@@ -1192,11 +1207,11 @@ module Slack
           'restricted_action_read_only_channel' => RestrictedActionReadOnlyChannel,
           'restricted_action_thread_locked' => RestrictedActionThreadLocked,
           'restricted_action_thread_only_channel' => RestrictedActionThreadOnlyChannel,
+          'restricted_plan_level' => RestrictedPlanLevel,
           'restricted_too_many' => RestrictedTooMany,
           'retention_override_not_allowed' => RetentionOverrideNotAllowed,
           'sent_recently' => SentRecently,
           'service_unavailable' => ServiceUnavailable,
-          'session_expired' => SessionExpired,
           'session_invalidation_failed' => SessionInvalidationFailed,
           'session_not_found' => SessionNotFound,
           'session_reset_not_allowed' => SessionResetNotAllowed,
@@ -1225,6 +1240,7 @@ module Slack
           'team_not_found' => TeamNotFound,
           'team_not_on_enterprise' => TeamNotOnEnterprise,
           'team_quota_exceeded' => TeamQuotaExceeded,
+          'team_tier_cannot_create_channel_canvases' => TeamTierCannotCreateChannelCanvases,
           'teams_not_found' => TeamsNotFound,
           'template_not_visible' => TemplateNotVisible,
           'thread_locked' => ThreadLocked,
@@ -1289,9 +1305,11 @@ module Slack
           'user_already_deleted' => UserAlreadyDeleted,
           'user_already_team_member' => UserAlreadyTeamMember,
           'user_cannot_create_channel' => UserCannotCreateChannel,
+          'user_cannot_manage_public_channels' => UserCannotManagePublicChannels,
           'user_cannot_manage_workspace' => UserCannotManageWorkspace,
           'user_disabled' => UserDisabled,
           'user_does_not_own_channel' => UserDoesNotOwnChannel,
+          'user_email_unverified' => UserEmailUnverified,
           'user_is_already_deleted' => UserIsAlreadyDeleted,
           'user_is_bot' => UserIsBot,
           'user_is_not_a_guest' => UserIsNotAGuest,