+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 2.4.1 (Next)
 
+* [#525](https://github.com/slack-ruby/slack-ruby-client/pull/525): Exclude spec files from gem package - [@amatsuda](https://github.com/amatsuda).
 * Your contribution here.
 
 ### 2.4.0 (2024/07/14)