+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * [#528](https://github.com/slack-ruby/slack-ruby-client/pull/528): Don't treat asterisks in list items as italic in markdown - [@rspeicher](https://github.com/rspeicher).
 * [#530](https://github.com/slack-ruby/slack-ruby-client/pull/530): Specify dependency on logger - [@rwstauner](https://github.com/rwstauner).
 * [#533](https://github.com/slack-ruby/slack-ruby-client/pull/533): Added `files_upload_external` - [@gregsaab](https://github.com/gregsaab).
+* [#534](https://github.com/slack-ruby/slack-ruby-client/pull/534): Rename `files_upload_external` to `files_uoload_v2` - [@gregsaab](https://github.com/gregsaab).
 * Your contribution here.
 
 ### 2.4.0 (2024/07/14)